+++ docs/versionhistory.rst
@@ -9,6 +9,8 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 - Changed all functions currently raising the private ``NoCurrentAsyncBackend``
   exception (since v4.12.0) to instead raise the public ``NoEventLoopError`` exception
   (`#1048 <https://github.com/agronholm/anyio/issues/1048>`_)
+- Fixed ``anyio.functools.lru_cache`` not working with instance methods
+  (`#1042 <https://github.com/agronholm/anyio/issues/1042>`_)
 
 **4.12.0**